removing useless implementation classes after realizing they are already implem
I was confused with how the elemental package worked and thought we
needed to implement the client side for each of the interfaces.  After
looking more I realized that the interfaces exists for us to use in
Java while the client side is already implemented.  Hopefully I am
understanding this now.